Final da aula 41

+++ app/Http/Controllers/UserInfoController.php
@@ -14,7 +14,7 @@ class UserInfoController extends Controller
      * de forma inicial.
      */
     public function __construct(){
-        $this->middleware('auth:web');
+        //$this->middleware('auth:web');
     }
 
     /**
@@ -36,13 +36,18 @@ class UserInfoController extends Controller
     public function create()
     {
         try {
-            $logged = Auth::user();
-            // procuro e vejo se existe a informção do usuário logado
-            $userInfo = UserInfo::find($logged->id);
-            if(isset($userInfo))
-                return view("UserInfo/show")->with("userInfo", $userInfo);
+            if(Auth::check())
+            {
+                $logged = Auth::user();
+                // procuro e vejo se existe a informção do usuário logado
+                $userInfo = UserInfo::find($logged->id);
+                if(isset($userInfo))
+                    return view("UserInfo/show")->with("userInfo", $userInfo);
+                else
+                    return view("UserInfo/create");
+            }
             else
-                return view("UserInfo/create");
+                return redirect()->route('login');
         } catch (\Throwable $th) {
             return view("UserInfo/create")->with("message", [$th->getMessage(), "danger"]);
         }
@@ -57,18 +62,23 @@ class UserInfoController extends Controller
     public function store(Request $request)
     {
         try{
-            // lembrar de dar o use Auth; (lá em cima)
-            $logged = Auth::user();
-            // lembrar de dar o use App\Models\UserInfo; (lá em cima)
-            $userInfo = new UserInfo();
-            $userInfo->Users_id = $logged->id;
-            // dados dentro do model = dados vindos da view
-            $userInfo->profileImg = $request->profileImg;
-            $userInfo->status = 'A';
-            $userInfo->dataNasc = $request->dataNasc;
-            $userInfo->genero = $request->genero;
-            $userInfo->save();
-            return view("UserInfo/show")->with("userInfo", $userInfo)->with("message", ["Informações cadastradas com sucesso", "success"]);
+            if(Auth::check())
+            {
+                // lembrar de dar o use Auth; (lá em cima)
+                $logged = Auth::user();
+                // lembrar de dar o use App\Models\UserInfo; (lá em cima)
+                $userInfo = new UserInfo();
+                $userInfo->Users_id = $logged->id;
+                // dados dentro do model = dados vindos da view
+                $userInfo->profileImg = $request->profileImg;
+                $userInfo->status = 'A';
+                $userInfo->dataNasc = $request->dataNasc;
+                $userInfo->genero = $request->genero;
+                $userInfo->save();
+                return view("UserInfo/show")->with("userInfo", $userInfo)->with("message", ["Informações cadastradas com sucesso", "success"]);
+            }
+            else
+                return redirect()->route('login');
         } catch(\Throwable $th){
             return view("UserInfo/create")->with("message", [$th->getMessage(), "danger"]);
         }
@@ -83,20 +93,25 @@ class UserInfoController extends Controller
     public function show($id)
     {
         try {
-            $logged = Auth::user();
-            // procuro e vejo se existe a informção do usuário logado
-            $userInfo = UserInfo::find($logged->id);
-            if(isset($userInfo))
-                if($id == $logged->id)
-                {
-                    return view("UserInfo/show")->with("userInfo", $userInfo);
-                }
+            if(Auth::check())
+            {
+                $logged = Auth::user();
+                // procuro e vejo se existe a informção do usuário logado
+                $userInfo = UserInfo::find($logged->id);
+                if(isset($userInfo))
+                    if($id == $logged->id)
+                    {
+                        return view("UserInfo/show")->with("userInfo", $userInfo);
+                    }
+                    else
+                    {
+                        return view("UserInfo/show")->with("userInfo", $userInfo)->with("message", ["Não é possível acessar informações de outros usuários", "warning"]);
+                    }
                 else
-                {
-                    return view("UserInfo/show")->with("userInfo", $userInfo)->with("message", ["Não é possível acessar informações de outros usuários", "warning"]);
-                }
+                    return view("UserInfo/create")->with("message", ["O usuário não possui informações adicionais cadastradas", "warning"]);
+            } 
             else
-                return view("UserInfo/create")->with("message", ["O usuário não possui informações adicionais cadastradas", "warning"]);
+                return redirect()->route('login');
         } catch (\Throwable $th) {
             return view("UserInfo/create")->with("message", [$th->getMessage(), "danger"]);
         }
@@ -111,23 +126,28 @@ class UserInfoController extends Controller
     public function edit($id)
     {
         try {
-            $logged = Auth::user();
-            // procuro e vejo se existe a informção do usuário logado
-            $userInfo = UserInfo::find($logged->id);
-            if(isset($userInfo))
+            if(Auth::check())
             {
-                if($id == $logged->id)
+                $logged = Auth::user();
+                // procuro e vejo se existe a informção do usuário logado
+                $userInfo = UserInfo::find($logged->id);
+                if(isset($userInfo))
                 {
-                    return view("UserInfo/edit")->with("userInfo", $userInfo);
+                    if($id == $logged->id)
+                    {
+                        return view("UserInfo/edit")->with("userInfo", $userInfo);
+                    }
+                    else {
+                        return view("UserInfo/show")->with("userInfo", $userInfo)->with("message", ["Não é possível acessar informações de outros usuários", "warning"]);
+                    }
                 }
-                else {
-                    return view("UserInfo/show")->with("userInfo", $userInfo)->with("message", ["Não é possível acessar informações de outros usuários", "warning"]);
+                else
+                {
+                    return view("UserInfo/create")->with("message", ["O usuário não possui informações adicionais cadastradas", "warning"]);
                 }
             }
             else
-            {
-                return view("UserInfo/create")->with("message", ["O usuário não possui informações adicionais cadastradas", "warning"]);
-            }
+                return redirect()->route('login');
         } catch (\Throwable $th) {
             return view("UserInfo/create")->with("message", [$th->getMessage(), "danger"]);
         }
@@ -144,22 +164,27 @@ class UserInfoController extends Controller
     {
         $userInfo = new UserInfo();
         try {
-            $logged = Auth::user();
-            $userInfo = UserInfo::find($logged->id);
-            if(isset($userInfo)) {
-                if($id == $logged->id){
-                    $userInfo->profileImg = $request->profileImg;
-                    $userInfo->genero = $request->genero;
-                    $userInfo->dataNasc = $request->dataNasc;
-                    $userInfo->update();
-                    return view("UserInfo/show")->with("userInfo", $userInfo)->with("message", ["Informações atualizadas com sucesso", "success"]);
-                } else {
-                    return view("UserInfo/show")->with("userInfo", $userInfo)->with("message", ["Não é possível acessar informações de outros usuários", "warning"]);
+            if(Auth::check())
+            {
+                $logged = Auth::user();
+                $userInfo = UserInfo::find($logged->id);
+                if(isset($userInfo)) {
+                    if($id == $logged->id){
+                        $userInfo->profileImg = $request->profileImg;
+                        $userInfo->genero = $request->genero;
+                        $userInfo->dataNasc = $request->dataNasc;
+                        $userInfo->update();
+                        return view("UserInfo/show")->with("userInfo", $userInfo)->with("message", ["Informações atualizadas com sucesso", "success"]);
+                    } else {
+                        return view("UserInfo/show")->with("userInfo", $userInfo)->with("message", ["Não é possível acessar informações de outros usuários", "warning"]);
+                    }
+                }
+                else {
+                    return view("UserInfo/create")->with("message", ["O usuário não possui informações adicionais cadastradas", "warning"]);
                 }
             }
-            else {
-                return view("UserInfo/create")->with("message", ["O usuário não possui informações adicionais cadastradas", "warning"]);
-            }
+            else
+                return redirect()->route('login');
         } catch (\Throwable $th) {
             if($userInfo->Users_id == null)
                 return view("UserInfo/create")->with("message", [$th->getMessage(), "danger"]);
@@ -176,11 +201,32 @@ class UserInfoController extends Controller
      */
     public function destroy($id)
     {
-        // procuro e vejo se existe a informção do usuário logado
-        $userInfo = UserInfo::find(1);
-        if(isset($userInfo)){
-            $userInfo->delete();
+        try {
+            if(Auth::check())
+            {
+                $logged = Auth::user();
+                // procuro e vejo se existe a informção do usuário logado
+                $userInfo = UserInfo::find($logged->id);
+                if(isset($userInfo)){
+                    // Estou tentando remover uma informação minha?
+                    if($id == $userInfo->Users_id)
+                    {
+                        // Caso a informação adicional exista e pertença ao usuário
+                        $userInfo->delete();
+                        return view("UserInfo/create")->with("message", ["Os dados foram removidos com sucesso", "success"]);
+                    } else {
+                        // Caso a informação adicional exista e não pertença ao usuário
+                        return view("UserInfo/show")->with("userInfo", $userInfo)->with("message", ["Não é possível acessar informações de outros usuários", "warning"]);
+                    }
+                } else {
+                    // Caso a informação adicional não exista
+                    return view("UserInfo/create")->with("message", ["O usuário não possui informações adicionais cadastradas", "warning"]);
+                }
+            } else {
+                return redirect()->route('login');
+            }
+        } catch (\Throwable $th) {
+            return view("UserInfo/create")->with("message", [$th->getMessage(), "danger"]);
         }
-        return view("UserInfo/create");
     }
 }
